move homepage component to container
